Rename entrypoint => servers
