+++ Makefile
@@ -3,10 +3,10 @@ start-deps:
 	direnv reload
 
 start: start-deps
-	. ./.envrc && yarn tsnd --respawn src/entrypoint/graphql-core-server.ts | yarn pino-pretty -c -l
+	. ./.envrc && yarn tsnd --respawn src/servers/graphql-core-server.ts | yarn pino-pretty -c -l
 
 start-admin: start-deps
-	. ./.envrc && yarn tsnd --respawn src/entrypoint/graphql-admin-server.ts | yarn pino-pretty -c -l
+	. ./.envrc && yarn tsnd --respawn src/servers/graphql-admin-server.ts | yarn pino-pretty -c -l
 
 watch:
 	yarn nodemon -V -e ts,graphql -w ./src -x make start
