+++ package.json
@@ -7,8 +7,8 @@
     "prettier-check": "prettier --check .",
     "prettier-write": "prettier --write .",
     "build": "tsc -p tsconfig-build.json --skipLibCheck && cp -R src/protos lib && cp src/schema.graphql lib",
-    "start": "yarn build && source ./scripts/export-local.sh && node lib/entrypoint/graphql-core-server.js | pino-pretty -c -l",
-    "trigger": "yarn build && source ./scripts/export-local.sh && node lib/entrypoint/trigger.js | pino-pretty -c -l",
+    "start": "yarn build && source ./scripts/export-local.sh && node lib/servers/graphql-core-server.js | pino-pretty -c -l",
+    "trigger": "yarn build && source ./scripts/export-local.sh && node lib/servers/trigger.js | pino-pretty -c -l",
     "watch": "nodemon -V -e ts,graphql -w ./src -x yarn run start",
     "watch-trigger": "nodemon -V -e ts,graphql -w ./src -x yarn trigger",
     "kill-graphql": "kill $(lsof -t -i:4000)",
@@ -18,8 +18,8 @@
     "forward-mongodb": "kubectl port-forward $(kubectl get pods -l app.kubernetes.io/name=mongodb | awk '/mongo/ {print $1}') 27017",
     "restart-charts": "yarn use-minikube && yarn uninstall-charts || : && sleep 5 && yarn run add-charts",
     "use-minikube": "kubectl config use-context minikube",
-    "cron": "yarn build && source scripts/export-local.sh && node src/entrypoint/cron.js",
-    "exporter": "yarn build && node lib/entrypoint/exporter.js",
+    "cron": "yarn build && source scripts/export-local.sh && node src/servers/cron.js",
+    "exporter": "yarn build && node lib/servers/exporter.js",
     "lnd-connection": "source scripts/export-local.sh && echo $MACAROON | base64 -d > ./macaroon && echo $TLS | base64 -d > ./tls.cert && lncli --network regtest --macaroonpath=./macaroon --tlscertpath=./tls.cert --rpcserver=${LNDDNS}:${LNDRPCPORT} getinfo",
     "lncli": "source scripts/export-local.sh && echo $MACAROON | base64 -d > ./macaroon && echo $TLS | base64 -d > ./tls.cert && lncli --network regtest --macaroonpath=./macaroon --tlscertpath=./tls.cert --rpcserver=${LNDDNS}:${LNDRPCPORT}",
     "postman": "yarn node src/postman/newman.ts",
