+++ test/integration/entrypoint/graphql.spec.ts
@@ -2,7 +2,7 @@ import { redis } from "src/redis"
 import { sleep } from "src/utils"
 import { yamlConfig } from "src/config"
 import { createTestClient } from "apollo-server-testing"
-import { startApolloServerForSchema } from "src/entrypoint/graphql-core-server"
+import { startApolloServerForSchema } from "src/servers/graphql-core-server"
 
 jest.mock("src/realtime-price", () => require("test/mocks/realtime-price"))
 jest.mock("src/phone-provider", () => require("test/mocks/phone-provider"))
